WebView/requirements.txt
pywebview
pywifi
asyncio
websockets
